Pdf png bug

+++ server/db/slides.db.js
@@ -1,10 +1,19 @@
 var fs = require('fs');
-var scissors = require('scissors');
+var path = require('path');
 var Duplex = require('stream').Duplex;
 var inspect = require('eyes').inspector({
   maxLength: 20000
 });
 var pdf_extract = require('pdf-extract');
+var pdf2img = require('pdf2img');
+pdf2img.setOptions({
+  type: 'png',
+  size: 1024,
+  density: 400,
+  outputdir: __dirname + path.sep + 'pdf2img_output',
+  outputname: 'test', // output file name, dafault null (if null given, then it will create image name same as input name)
+  page: null // convert selected page, default null (if null given, then it will convert all pages)
+});
 
 var Slide = require('../models/slide');
 
@@ -26,38 +35,21 @@ function createFromLecture(lecture) {
     } else {
       /*
        * 1. Store the PDF presentation in a temp file on disk
-       * 2. Extract text from all slides and take keep track of paths of single page PDF
-       *    left over by pdf-extract processor
-       * 3. Convert each single sided PDF file into a PNG image
+       * 2. Extract text from all slides
+       * 3. Extract images from all slides
        * 4. Group the extracted data into Slide models
        * 5. Bulk insert all into the database
+       * 6. Clean up the temp PDF file and resolve
        */
       fs.writeFile(TEMP_PRESENTATION, lecture.file, 'binary', (err) => {
         if (err) {
           reject(err);
         } else {
-          extractSlidesTextArray(TEMP_PRESENTATION).then((slidesArr) => {
+          extractSlidesTextArray(TEMP_PRESENTATION).then((slidesTextArr) => {
             // 3
-            var promisesQueue = [];
-            for (var i = 0; i < slidesArr.length; i++) {
-              promisesQueue.push(extractImageFromSlide(slidesArr[i]));
-            }
-            Promise.all(promisesQueue).then((images) => {
-              console.log("[INFO] all extract image promises resolved");
-              var slides = [];
-              for (var j = 0; j < slidesArr.length; j++) {
-                slides.push(new Slide({
-                  lectureId: lecture._id,
-                  image: images[j],
-                  text: slidesArr[j].text,
-                  isQuiz: false,
-                  slideNumber: j + 1
-                }));
-              }
-              Slide.insertMany(slides).then((docs) => {
-                console.log("Inserted " + docs.length + " slides");
-                resolve();
-              }).catch((err) => {
+            pdf2img.convert(TEMP_PRESENTATION, (err, info) => {
+              if (err || info === undefined || !info.hasOwnProperty('result') || info['result'] !== 'success' ||
+                !info.hasOwnProperty('message')) {
                 // clean up the temp file
                 fs.unlink(TEMP_PRESENTATION, (err2) => {
                   if (err2) {
@@ -65,15 +57,53 @@ function createFromLecture(lecture) {
                   }
                   reject(err);
                 });
-              });
-            }).catch((err) => {
-              // clean up the temp file
-              fs.unlink(TEMP_PRESENTATION, (err2) => {
-                if (err2) {
-                  console.error(err2);
+              } else {
+                var promisesQueue = [];
+                var slidesImagesArr = info['message'];
+                for (var i = 0; i < slidesImagesArr.length; i++) {
+                  promisesQueue.push(convertImageToBuffer(slidesImagesArr[i]));
                 }
-                reject(err);
-              });
+                Promise.all(promisesQueue).then((images) => {
+                  // 4
+                  console.log("[INFO] all extract image promises resolved");
+                  var slides = [];
+                  for (var j = 0; j < slidesTextArr.length; j++) {
+                    slides.push(new Slide({
+                      lectureId: lecture._id,
+                      image: images[j],
+                      text: slidesTextArr[j],
+                      isQuiz: false,
+                      slideNumber: j + 1
+                    }));
+                  }
+                  Slide.insertMany(slides).then((docs) => {
+                    console.log("Inserted " + docs.length + " slides");
+                    // clean up the temp file from disk and resolve
+                    fs.unlink(TEMP_PRESENTATION, (err2) => {
+                      if (err2) {
+                        console.error(err2);
+                      }
+                      resolve();
+                    });
+                  }).catch((err) => {
+                    // clean up the temp file
+                    fs.unlink(TEMP_PRESENTATION, (err2) => {
+                      if (err2) {
+                        console.error(err2);
+                      }
+                      reject(err);
+                    });
+                  });
+                }).catch((err) => {
+                  // clean up the temp file
+                  fs.unlink(TEMP_PRESENTATION, (err2) => {
+                    if (err2) {
+                      console.error(err2);
+                    }
+                    reject(err);
+                  });
+                });
+              }
             });
           }).catch((err) => {
             // clean up the temp file
@@ -158,12 +188,9 @@ function extractSlidesTextArray(filePath) {
   return new Promise((resolve, reject) => {
     /*
      * 1. Use pdf_extract to extract text from each slide
-     * 2. Keep track of the paths to pdf single page file leftovers and append
-     *    them to the output array for future PDF -> PNG conversion
      */
     var options = {
       type: 'text',
-      clean: false // keeps the single page pdfs on disk after processor completes
     };
     var processor = pdf_extract(filePath, options, (err) => {
       if (err) {
@@ -172,10 +199,7 @@ function extractSlidesTextArray(filePath) {
         processor.on('complete', (data) => {
           var slides = [];
           for (var i = 0; i < data.text_pages.length; i++) {
-            slides.push({
-              text: data.text_pages[i],
-              slidePath: data.single_page_pdf_file_paths[i]
-            });
+            slides.push(data.text_pages[i]);
           }
           resolve(slides);
         });
@@ -187,34 +211,32 @@ function extractSlidesTextArray(filePath) {
   });
 }
 
-function extractImageFromSlide(slide) {
+function convertImageToBuffer(slideImageOnDisk) {
   return new Promise((resolve, reject) => {
-    /*
-     * 1. Extract the png from the single page PDF provided
-     * 2. Clean up the slide from disk before resolving the image back
-     */
-    var pngStream = scissors(slide.slidePath).pages(1).pngStream(300).on('error', (e) => reject(e));
-    streamToBuffer(pngStream).then((img) => {
-      fs.unlink(slide.slidePath, (err) => {
-        // does not matter
+    if (slideImageOnDisk === null) {
+      reject("Parameter slideImageOnDisk was null");
+    } else if (slideImageOnDisk === undefined) {
+      reject("Parameter slideImageOnDisk was undefined");
+    } else if (!slideImageOnDisk.hasOwnProperty("path")) {
+      reject("Parameter slideImageOnDisk is malformed and missed the path property");
+    } else {
+      /*
+       * 1. Read the image from the path provided into a Buffer
+       * 2. Remove the file from disk
+       * 3. Resolve the in-memory image
+       */
+      fs.readFile(slideImageOnDisk['path'], (err, img) => {
         if (err) {
-          console.warn("[WARN]: " + err);
+          reject(err);
+        } else {
+          fs.unlink(slideImageOnDisk['path'], (err2) => {
+            if (err2) {
+              console.warn("[WARN] Failed to remove the temp image from disk: " + err2);
+            }
+            resolve(img);
+          });
         }
-        resolve(img);
       });
-    }).catch((err) => {
-      console.error("[ERR] extractImageFromSlide failed: " + err);
-      reject(err);
-    });
-  });
-}
-
-// http://derpturkey.com/buffer-to-stream-in-node/
-function streamToBuffer(stream) {
-  return new Promise((resolve, reject) => {
-    let buffers = [];
-    stream.on('error', reject);
-    stream.on('data', (data) => buffers.push(data));
-    stream.on('end', () => resolve(Buffer.concat(buffers)));
+    }
   });
 }
